feat: remove signing for factory
Initially it was added because we were recovering from the signature
to determine who the user was. But now the interface has changed
and we added a `user` address param to support arbitrary encoded
signatures for smart contract 1271 signatures.

With `user` parameter now available, we can directly call the
`executeHooks` method on the proxy after the deployment/initialization.