feat: COWShedProxy guaranteed to be owner controlled always
By using an immutable admin variable and a special function to set the
implementation directly in the proxy, we can guarantee that the user
will always be in control of proxy even if by some way some phising scam
got them to sign a hook that changed the implementation to something
else, etc.

But, because we ended up using an immutable variable, the automatic code
verification on etherscan will not work either. But it was decided it is
worth using an immutable vs a storage variable that needs to be checked
in each tx for alterations, 400+ gas every call.

A offchain bot can listen for `COWShedFactory::COWShedBuilt` events to
listen for newly deployed proxies and verify it for the user as a
service.