feat: remove admin storage slot from shed impl
The admin is already available as an immutable in the proxy. We can
expose the same from the proxy via a special view function. The view
function is also special cased and only returns the admin if the
msg.sender is the proxy itself. i.e. all the calls from within the
implementation will return the admin immutable. But any calls from any
other address will be transparently proxied to the implementation
contract.