More robust deterministic deployments (#28)
After some time passes, deterministic deployments are hard to replicate
using the tooling on a Foundry repo, as seen in the readme changes from
#27.

Because of a relatively minor security issue in the contracts (that will
be addressed in another PR), we'll need to redeploy the contracts. We
take this opportunity to hopefully make the experience of deploying to a
new chain more robust.

Precisely, what this PR does is the following:
- It sets Foundry stable v1.2.1 as the default version for this project.
- It extracts from the current deployment script the part that does the
actual deployment.
- It adds a test that verifies that the generated contract addresses
matches the addresses that are hardcoded in the test. Right now, these
addresses are meaningless; however, once we'll redeploy the contracts,
these hardcoded addresses will match the official addresses and can be
used to determine if a commit breaks deterministic deployments.
- It adds the Solidity standard JSON input for both contract. This can
be used on most block explorer to manually verify the contract code on
the web interface.
- It adds a script that automatically generates the standard JSON input.
- It adds a CI job validating that the files match what is generated by
the script.
- It updates the readme with the process changes.
- My editor's autoformatting of the readme (on request I can undo this).

### How to test

This PR already adds some tests.
I tried the deployment script and the result can be found
[here](https://sepolia.etherscan.io/address/0x4dD02D5e51fAE5A59D5B780A9E61a33c31580A46#code).
There are minor changes to the code so that I could also try out
generating the JSON standard input and verifying through, which was
successful.
The new CI test is untested.