Isolate forked tests (#34)
All test of the shed depend on the current live mainnet state.
This is often undesirable: for example, testing the deployment script
fails if the contract has already been deployed on mainnet, as seen in
#33.

The changes in this PR create a clear split between local and forked
tests.
The only tests that have been made local-only are `Deploy.t` and
`LibAuthenticatedHooks.t` for simplicity and to keep the related changes
minimal.

### Details

Note that by how forking has been implemented, we can't directly deploy
contracts in the test contract body but any deployment must be moved to
the setup. `LibAuthenticatedHooksCalldataProxy` has been isolated to its
own file.

### How to test

See that all tests are included in CI. Eventually. merge this PR into
#33 and see that CI passes.

---------

Co-authored-by: Anxo Rodriguez <anxo@cow.fi>