Release v1.0.1 (#33)
Deploy v1.0.1 version, which includes mostly the fix of
https://github.com/cowdao-grants/cow-shed/pull/29

## Additional changes in the contract
Increase version and move it to a const:
https://github.com/cowdao-grants/cow-shed/pull/33/commits/f1805cd624d7b71294f6cac9f3c991de1e477c17

No strong reason to do this, but since the code is changed, we felt like
versioning the scheme to signal is not the same version. Technically was
not needed.

A v1.1.0 version is coming, so its kind of nice to be able to read the
version from on-chain. This is why also the version was moved to a
public constant.

## Deployment in all networks
It was deployed in all the networks and updated the `networks.json` (see
https://github.com/cowdao-grants/cow-shed/blob/release-v1.0.1/networks.json)

## Misc: Changes unrelated to the PR 
If I may, let me do some small changes
### Remove `Pre-requisites` section in the README
I don't think is needed anymore for these contracts since they are
deployed with:
```
evm_version = "prague" # https://docs.soliditylang.org/en/latest/using-the-compiler.html#target-options
bytecode_hash = "none"
cbor_metadata = false
```
So the address will continue to be deterministic in the future (I hope).

## Renamed broadcast/Deploy.s.sold
It was renamed in a past PR to DeployAndRecord, but the directory was
not renamed. I just `git mv` it

## Remove weird $ in commands
Yes, annoys me a lot :) 
Why would foundry add a $ in the `shell` script? 
Its silly! many markdown readers allow you to copy the content so you
can paste in your terminal, and the command that is copied won't work.

I sneakily left this change camouflaged in the last point of the misc
section, hoping no one reads it.

# Test
Make sure the contracts are correctly deployed and verified and
networks.json reflects the new deployments:


https://github.com/cowdao-grants/cow-shed/blob/release-v1.0.1/networks.json