Enforce formatting rules (#37)
We had many PRs that introduced small logic changes but many changes in
the code. This is due to inconsistent code formatting between the
existing code in the repo and the people who wrote the PR.

This PR tries to avoid this issue in the future. Namely:
- It runs `forge fmt` on the entire codebase.
- It introduces a new CI step that confirms that the commit formatting
is consistent with the repo rules.
- (Opinionated.) It enforces sorted imports.
- It manually removes the newlines introduced by `forge fmt` because of
[this bug](https://github.com/foundry-rs/foundry/issues/7944).
- Since I was making changes to the imports anyway, I also removed
unused imports (found by `forge lint` on nightly).

### How to test

Check that there's a new CI step and it passes.