Expose proxy init code as a factory read method (#35)
This PR adds a view method named `PROXY_CREATION_CODE` that returns the
init code of the CoW Shed proxy.
This can be used by integrator to easily recover this information if the
contract is redeployed, specifically if they want to generate the proxy
address deterministically without requiring a call to the contract.

Note: I wanted to remove `deploymentAddresses.json` and
`DeployAndRecord` but I've noticed it's still used in
`examples/common.ts` for what appears to be code for manual testing. I
also fixed that file since it was using the deployment script that
doesn't generate the file. On the other hand, now the main deployment
procedure doesn't need to use `DeployAndRecord` but can just use
`Deploy` as the only missing information (the init code) is now easily
accessible.

### How to test

New unit tests. I added a test that checks that `proxyOf` indeed returns
the address of the initialized proxy: this was implicitly tested in
other tests, but now this is made explicit.