Allow network json to be manually generated (#38)
This PR tries to address this comment
https://github.com/cowdao-grants/cow-shed/pull/36#pullrequestreview-3052452531
differently.

For me, its super useful to generate the `networks.json` file
automatically using the script, and its a pity that the issue deploying
in Lens makes it sothat new networks we need to do manually.

## Changes
This PR introduces a new file `networks-manual.json` with the same shape
as `networks.json`. If we have a network that we deployed manually, we
add it there, the rest we will have the broadcast artifacts.

The script merges the result.

## Test
Regenerate the `networks.json` file and make sure there's no git changes
after that.
```
bash dev/generate-networks-file.sh > networks.json
```

---------

Co-authored-by: Federico Giacon <58218759+fedgiac@users.noreply.github.com>