Add presign flow (#30)
Cow-shed provides 2 signing schemes: ERC712 signature and 1271. 

This PR adds the possibility of pre-authorise (pre-sign) some nonce, so
it can be executed by any account after that point.

The PR has the following changes

## Motivation
This simplifies a lot the combination of presign flow for CoW Protocol
and the pre-authorization of hooks.
One example is in the Safe flow in CoW Swap. If we used ERC712 we would
require to wait until the latest signer signed before we can compose
fully the hook with the all the required signatures.

If we have this flow, now we can bundle in the same safe tx the presign
of the hook, and the order together. This will simplify adding things
like swap and bridge to CoW Swap.

## Keeps track of signed nonces
The state has now a `Bitmap signedNonces`. 

This will tell us if the nonce has been signed or not. 


## Pre-sign method
`function preSignHook(Call[] calldata calls, bytes32 nonce, uint256
deadline, bool signed) external onlyAdmin`

Added a new method in cow-shed for pre-signing.
This method doesn't ask for any nonce because it uses a convention for
the generation of the nonce. This is done to be able to derive the nonce
deterministically from the calls+deadline.

The same method allows to both presign and revoke the signature (with
the`signed` param)

## Expose the pre-signed state
Exposes a view function to check if the hook has been pre-signed

```solidity
isPreSignedHook(Call[] calldata calls, bytes32 nonce, uint256 deadline) external view returns (bool)
```

## Execute pre-signed hooks
`executePreSignedHooks(Call[] calldata calls, bytes32 nonce, uint256
deadline)`

Allows you to execute some hooks that have been previously authorised. 
The implementation is very straight forward. It use the same `nonce`
generation convention asn the pre-sign method to get the nonce, then it
checks in the bitmap if it has been signed. If so, it will auth the
call, if not it will revert.

## Versioning
~I bumped the minor version of the contract. Initially I increased the
major one, but I think the contract is still backward compatible, it
just adds some methods and additional data. So using `1.1.0` (let me
know if you think it should be 2.0.0 because of the state struct
returning holding more data, but afait I consider backward compatible)~

I will bump the version when we do a deployment (as suggested in the
comments)


## Naming
I tried to make the new methods coherent with the other methods, but if
this was a new contract, I would not use the plural form in:

```solidity
function preSignHooks(Call[] calldata calls, bytes32 nonce, uint256 deadline, bool signed) external;
function isPreSignedHooks(Call[] calldata calls, bytes32 nonce, uint256 deadline) external view returns (bool);
function executePreSignedHooks(Call[] calldata calls, bytes32 nonce, uint256 deadline) external;
```

I did it because there's methods `executeHooks` and
`trustedExecuteHooks`, but imo, even if its an array of `Call[] calldata
calls` is a single hook. Maybe there should not even a reference to the
word hooks here, after all, the shed allows a user to execute some
calls, is not related to hooks at all (this doesn't need to be executed
in the context of CoW Protocol hooks)


## Test
Run tests. It should run the:
- happy path
- unhappy path
- revoking a pre-authorized call
- revoking when we didn't do the pre-authorized call
- Executing a pre-authorized hook twice
- Several tests to assert the correct behaviour of the `isPreSignedHook`

```bash
forge test --fork-url $RPC_URL --match-test testPreSignFlow
```

<img width="561" alt="image"
src="https://github.com/user-attachments/assets/b37d981f-ac2c-4fe5-b91f-eb0d7c0a81e8"
/>

---------

Co-authored-by: Federico Giacon <58218759+fedgiac@users.noreply.github.com>