Test that authentication fails when signatures are invalid (#43)
When using the output of ` forge coverage --report lcov --no-match-test
testMatchesOfficialAddresses`, I noticed that we don't have a test
checking that if the signatures are invalid then the transaction isn't
executed.

This PR adds some tests for covering the simplest revert conditions.

### How to test

CI passes. Relevant lines are covered when running the command above.