Opt-in presign flow + move state out  (#41)
Closes #39

# Description
Makes pre-sign flow (implemented in #30) an opt-in feature.

## Context
This PR was created as a suggestion from @fedgiac to address this:
> it becomes harder to validate the state of a shed: if the shed is ever
compromised, then an attacker could presign a transaction for the future
that can be executed on an apparently clean shed to retake the ownership
of the shed

For full context see #39 and #30

## How it works
- The pre-signatures are not stored in CoW Shed storage anymore, instead
they use a different contract
- When the shed is initialized, the storage is set to zero (meaning that
the feature is disabled)
- Any user can enable the feature by deploying a PreSignStateStorage and
registering that address on the shed

If something ever happens, the user can redeploy that storage contract
again and point to that, effectively clearing the state.

# Deploying PreSignStateStorage
For convenience, this PR gives 3 options to create the contract.

In the factory, you can deploy the contract for a given cowShed or the
owner respectively
```solidity
deployPreSignStateStorage(address cowShed) external returns (address)
deployPreSignStateStorageForUser(address user) external returns (address)
```

Once you've deployed the contract, you can set it to the relevant shed
using:
```solidity
setPreSignStorage(address storageContract) external onlyAdmin
```

These 2 factory methods are convenient for deploying the storage. 
However, most users will just need to have a easy way to opt-in to the
feature, this is why I added an initialize method that sets the storage
address to a brand new contract in a single call:

```solidity
initializePreSignStorage() external onlyAdmin;
```

This method will come handy for things like:
- Simplify the flow to opt-in to pre-sign flow (1 step)
- Give the user a easy way to revoke all pre-signatures 

## Revoking all signatures

It can be done in two ways:

1. `shed.initializePreSignStorage()`: Re-asign the storage to a new
contract, effectively clearing all previously approved calls
2. `shed.setPreSignStorage(address(0))`: Technically this invalidates
all previously approved calls, and also opts-out from the pre-sign flow.
After this an attempt of pre-signing should revert


## Swap and bridge flow (Safe support)
In this section, I detail how I'm expecting this to be used for Swap and
Bridge Safe flow. This is a very relevant example, but the same flow
should work with any hook-based flow for any smart contract wallet.

When placing the order, we need to create a safe tx that bundles the
following calls:
- approve sell token: `sellToken.approve(vaultRelayer, sellAmount)`
- presign the orderUid in settlement:
`settlement.setPresignature(orderUid)`
- deploy cow-shed: `factory.initializeProxy(user)`
- initialize pre-sign storage: `shed.initializePreSignStorage()`
- pre-sign hooks: `shed.preSignHooks(calls, nonce, deadline, true)`

The rest of the flow should be very similar to the EOAs. Although there
are some differences:
* In this flow, by the time the hook is executed, we know cow-shed has
already been initialized. This is because, the intialization happens at
order creation. This means that we can invoke the shed directly. No need
to call it through the factory (as CoW Swap does for EOA flows)
* The method we need to call is not the same. EOA uses `executeHooks`,
which require a ERC-712 signature. In this flow our hook calldata would
be already pre-authorized, so we don't need a signature. Therefore, the
hook encodes a call to `executePreSignedHooks(calls, nonce, deadline)`
instead.

## How to review this PR
- Main contract is `PreSignStateStorage`
- Check the new methods in `COWShed` and `COWShedFactory`
- Check the new tests in `COWShedPreSign.t.sol` and
`COWShedFactory.t.sol`

## Test
This PR adds many tests for all the new methods introduced, mainly in
`COWShedPreSign.t.sol` and `COWShedFactory.t.sol`:

```bash
forge test
```

<img width="1073" height="883" alt="image"
src="https://github.com/user-attachments/assets/d62d16d2-88c3-4a4b-aa24-450fba0a0ec9"
/>

---------

Co-authored-by: Federico Giacon <58218759+fedgiac@users.noreply.github.com>