Bump version to 2.0.0 (#51)
We already bumped version to 1.1.0 in
767048f9d752dfb25808139283f009741c60263c, but after some discussion we
decided that because of the removal of the ENS feature the contract code
makes a breaking change and deserves its own major version bump
according to semantic versioning (different function signatures in the
initialization). Version 1.1.0 would then officially "disappear" as it
would be the same as version 2.0.0.

<details><summary>More points about versioning copied from
@anxolin</summary>

For the pre-sign and old flow flows, I would think this version should
increase only the "minor version" because

- Just adds new methods for the pre-sign
- Old flows works exactly as it used to
- Storage respect the same position for the data (it just includes
additional info at the end of the struct)

Regarding the new admin flow:

- I consider it backward compatible because you can still use the
trustedExecuteHooks as before (from the factory), it doesn't matter than
now you can also do it from the admin account

However, and here it comes.... I think we need to increase the MAJOR
version according to sem versioning because of the ENS change. External
function initialize changed its signature, and claimWithResolver is
gone.

</details>

### How to test

As in 767048f9d752dfb25808139283f009741c60263c: CI. No instances of
1.0.1 or 1.1.0 appear in the codebase anymore.