Add audit update for version 2.0.0 (#52)
The audit for the latest contract changes has been concluded.
This PR adds the audit update document to this repo.

This change is exactly the same as #50, except that the document refers
to version 2.0.0 instead of 1.1.0 (same logic, just different version
number, see bbfba8cc6ab0916b5eeb9fce940ca18cb7f13701).

I also took the liberty of renaming the audit files for clarity.

### How to test

Read the audit and make sure that all issues are fixed and that it
matches the code in this repo (notably the commit hashes).

If you already reviewed #50: you can just check with tools like
`diffpdf` that the only difference with the PDF from #50 is
`s/1.1.0/2.0.0/` as well as an updated commit hash.