Implement composable cow for cow-shed (#53)
This experimental PR aims to add composable cow support to cow-shed (see
`COWShedForComposableCoW.sol`).

The goal is to find a feasible solution based on cow-shed. 

## Why do it in cow-shed?

- The contract is simple compared to a full multisig, but still lets an
EOA bundle and execute calls from this account.
- Users will likely need to perform operations such as creating or
canceling orders, which require authentication. The cow-shed already has
built-in authentication for this.
- It could also be useful to allow arbitrary contract calls from the
"trader account" (the shed). This would allow to do some fancy
programmatic orders - for example, an order that reads a value from an
external contract where any address can write to a mapping (msg.sender
-> value). This pattern can be very handy for certain programmatic
strategies.
- Why adding it in cow-shed? Why not? To support flexibility, I created
the shed as an independent contract. This way, users can opt in by using
the new contract with extended functionality, or opt out by sticking to
the plain cow-shed contract. So should not bother swap-n-bridge
- We already have tooling in the SDK and CoW Swap to work with the shed,
plus UIs for the recovery of tokens and keeping track of things
(balances, etc). So less FE work by not introducing another new
contract!

## End to end flow
Let's imagine an EOA wants to make a TWAP. I will explain how this could
work in a simple implementation. Some small variations of this flow can
be made more capital efficient, but lets keep it simple for now.

The user wants:
* Sell `10K sellToken` in `4 parts` for `buyToken`

### 1️⃣ First, we prepare 2 calls (no signing needed)
- **CALL1**: Encode call to enable sellToken for trading:
`sellToken.approve(fullAmountOfTrade)`
- **CALL2** Encode call to `ComposableCow.create()` with
    - `handler`: TWAP address
    - `salt`: whatever
- `staticInput`: The encoded TWAP order to sell in `4 parts` `10K` of
`sellToken` for `buytoken`


### 2️⃣ Then the user signs the calls
Encode a call to the factory method `executeHooks` with the two calls,
and sign with an EOA
- **SIGNED_HOOK**: `cowShedFactory.executeHooks([ CALL1, CALL2 ])`

We now have the signed hook that creates the TWAP!


### 3️⃣ Lastly, the user signs the order and post it

- **Sell**: `sellToken`
- **Buy**: `sellToken` (no, is not a mistake, also we depend on sell=buy
using this approach)
- **receiver**: cowShedForComposableCow (this new shed that has support
for composable cow)
- **appData**
    - postHook: `SIGNED_HOOK ` (from the previous step)


### 4️⃣ When the order is executed
 - cow-shed receives the funds (`10K sellToken`)
 - The post hook
    - enables the `sellToken` for trading in CoW Protocol
    - Creates the TWAP

### 5️⃣ Watch tower picks the order
- The TWAP contract will yield the first part
- Watch Tower will post the order to the API

### 6️⃣ A settlement includes the order
- `isValidSignature` is invoked with the first part of the TWAP on the
cow-shed
- The cow-shed delegates the call to composable cow
- which delegates to TWAP
- which accepts signing the order
- Because the funds are in the owner contract, and the approval is in
place --> 🎉 The trade happens


# Challenges and posible solutions

## Dependency with sell=buy
Not implemented yet in the backend, but this will be another incentive
that proves how useful it is. This allows seemesly to do a programatic
order GASLESS!

But are we really blocked by backend? NO. I think FE can start working
even before, because we can test this just using a different `sellToken`
than the `buy token`. It will still work!

Also, we can reuse the same recovery tool if funds get stuck, so less FE
work!

## Its capital inefficent
For TWAP sending all parts at once might be OK if you do a TWAP for 24h
or 1 week, but longer periods might need another approach. Same for stop
loss for example.

This is a solvable problem by:
- The EOA sets an approval to cow-shed so it can pull the funds for the
`sellToken`
- We add an additional pre-hook that moves the funds just in time

## Possible collisions
If the funds are not used right away and stay for some period in the
shed, depending on how we implement the things there could be
collisions. For TWAP is not a problem (I can have 2 open TWAP orders
using the same shed), but some orders might have issues with this.

If this is the case, we could add a salt to the shed to keep the
independence.

## Amount of the trade
If done as suggested. We can't predict exactly how many tokens we will
get as buy token, so `fullAmountOfTrade` is unknown.

There are multiple possible solutions:
- Buy order: So we buy exactly `fullAmountOfTrade`. This feels the most
natural for the twap and the simpler to implement
- Auxiliary contract for the post-hook: If we can't predict the amount
bought, we will need a contract that gets the balance in the shed +
approves that balance + creates TWAP

## Deployment
Only Gnosis Chain network is deployed. Rest of networks will be a follow
up PR after the unit tests and applying all the nits.


## Out of scope (different PRs)
Next steps
- [ ] I could deploy the contracts, and make a PoC purely using the SDK 
- [ ] We could do a unit test
- [ ] Deploy in all networks
- [ ] Solve the capital inefficiency: JIT pulling of funds
- [ ] Re-think how combined storage of the 2 contract (in this PR one
uses a specific slot for cow-shed state, but the composable cow is
written directly in the first slots)